fix some formating; create a monkey amnimation so i dont loose focus when running the script and move reporter proxy to a reusable function in testing-utils.mjs
